start arsing about with templates
