+++ internal/router/template.go
@@ -19,27 +19,12 @@ package router
 
 import (
 	"fmt"
-	"html/template"
 	"os"
 	"path/filepath"
-	"strings"
-	"time"
 
 	"github.com/gin-gonic/gin"
-	apimodel "github.com/superseriousbusiness/gotosocial/internal/api/model"
 	"github.com/superseriousbusiness/gotosocial/internal/config"
-	"github.com/superseriousbusiness/gotosocial/internal/log"
-	"github.com/superseriousbusiness/gotosocial/internal/text"
-	"github.com/superseriousbusiness/gotosocial/internal/util"
-)
-
-const (
-	justTime     = "15:04"
-	dateYear     = "Jan 02, 2006"
-	dateTime     = "Jan 02, 15:04"
-	dateYearTime = "Jan 02, 2006, 15:04"
-	monthYear    = "Jan, 2006"
-	badTimestamp = "bad timestamp"
+	"github.com/superseriousbusiness/gotosocial/internal/template"
 )
 
 // LoadTemplates loads html templates for use by the given engine
@@ -62,128 +47,6 @@ func LoadTemplates(engine *gin.Engine) error {
 	return nil
 }
 
-func oddOrEven(n int) string {
-	if n%2 == 0 {
-		return "even"
-	}
-	return "odd"
-}
-
-func escape(str string) template.HTML {
-	/* #nosec G203 */
-	return template.HTML(template.HTMLEscapeString(str))
-}
-
-func noescape(str string) template.HTML {
-	/* #nosec G203 */
-	return template.HTML(str)
-}
-
-func noescapeAttr(str string) template.HTMLAttr {
-	/* #nosec G203 */
-	return template.HTMLAttr(str)
-}
-
-func timestamp(stamp string) string {
-	t, err := util.ParseISO8601(stamp)
-	if err != nil {
-		log.Errorf(nil, "error parsing timestamp %s: %s", stamp, err)
-		return badTimestamp
-	}
-
-	t = t.Local()
-
-	tYear, tMonth, tDay := t.Date()
-	now := time.Now()
-	currentYear, currentMonth, currentDay := now.Date()
-
-	switch {
-	case tYear == currentYear && tMonth == currentMonth && tDay == currentDay:
-		return "Today, " + t.Format(justTime)
-	case tYear == currentYear:
-		return t.Format(dateTime)
-	default:
-		return t.Format(dateYear)
-	}
-}
-
-func timestampPrecise(stamp string) string {
-	t, err := util.ParseISO8601(stamp)
-	if err != nil {
-		log.Errorf(nil, "error parsing timestamp %s: %s", stamp, err)
-		return badTimestamp
-	}
-	return t.Local().Format(dateYearTime)
-}
-
-func timestampVague(stamp string) string {
-	t, err := util.ParseISO8601(stamp)
-	if err != nil {
-		log.Errorf(nil, "error parsing timestamp %s: %s", stamp, err)
-		return badTimestamp
-	}
-	return t.Format(monthYear)
-}
-
-type iconWithLabel struct {
-	faIcon string
-	label  string
-}
-
-func visibilityIcon(visibility apimodel.Visibility) template.HTML {
-	var icon iconWithLabel
-
-	switch visibility {
-	case apimodel.VisibilityPublic:
-		icon = iconWithLabel{"globe", "public"}
-	case apimodel.VisibilityUnlisted:
-		icon = iconWithLabel{"unlock", "unlisted"}
-	case apimodel.VisibilityPrivate:
-		icon = iconWithLabel{"lock", "private"}
-	case apimodel.VisibilityMutualsOnly:
-		icon = iconWithLabel{"handshake-o", "mutuals only"}
-	case apimodel.VisibilityDirect:
-		icon = iconWithLabel{"envelope", "direct"}
-	}
-
-	/* #nosec G203 */
-	return template.HTML(fmt.Sprintf(`<i aria-label="Visibility: %v" class="fa fa-%v"></i>`, icon.label, icon.faIcon))
-}
-
-// text is a template.HTML to affirm that the input of this function is already escaped
-func emojify(emojis []apimodel.Emoji, inputText template.HTML) template.HTML {
-	out := text.Emojify(emojis, string(inputText))
-
-	/* #nosec G203 */
-	// (this is escaped above)
-	return template.HTML(out)
-}
-
-func acctInstance(acct string) string {
-	parts := strings.Split(acct, "@")
-	if len(parts) > 1 {
-		return "@" + parts[1]
-	}
-
-	return ""
-}
-
-func increment(i int) int {
-	return i + 1
-}
-
 func LoadTemplateFunctions(engine *gin.Engine) {
-	engine.SetFuncMap(template.FuncMap{
-		"escape":           escape,
-		"noescape":         noescape,
-		"noescapeAttr":     noescapeAttr,
-		"oddOrEven":        oddOrEven,
-		"visibilityIcon":   visibilityIcon,
-		"timestamp":        timestamp,
-		"timestampVague":   timestampVague,
-		"timestampPrecise": timestampPrecise,
-		"emojify":          emojify,
-		"acctInstance":     acctInstance,
-		"increment":        increment,
-	})
+	engine.SetFuncMap(template.FuncMap)
 }
