+++ web/template/404.tmpl
@@ -19,21 +19,21 @@
 
 {{ template "header.tmpl" .}}
 <main>
-	<section>
-		<h1>404: Page Not Found</h1>
-		<p>
-			GoToSocial only serves Public statuses via the web.
-			If you reached this page by clicking on a status link,
-			it's possible that the status is not Public, has been
-			deleted by the author, you don't have permission to see
-			it, or it just doesn't exist at all.
-		</p>
-		<p>
-			If you believe this 404 was an error, you can contact
-			the instance admin. Provide them with the following request
-			Request ID: <code>{{.requestID}}</code>.
-		</p>
-	</section>
+    <section>
+        <h1>404: Page Not Found</h1>
+        <p>
+            GoToSocial only serves Public statuses via the web.
+            If you reached this page by clicking on a status link,
+            it's possible that the status is not Public, has been
+            deleted by the author, you don't have permission to see
+            it, or it just doesn't exist at all.
+        </p>
+        <p>
+            If you believe this 404 was an error, you can contact
+            the instance admin. Provide them with the following request
+            Request ID: <code>{{.requestID}}</code>.
+        </p>
+    </section>
 </main>
 
 {{ template "footer.tmpl" .}}
