+++ web/template/authorize.tmpl
@@ -17,26 +17,24 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */ -}}
 
-{{ template "header.tmpl" .}}
-    <main>
-        <form action="/oauth/authorize" method="POST">
-            <h1>Hi {{.user}}!</h1>
-            <p>
-              Application <b>{{.appname}}</b> 
-              {{if len .appwebsite | eq 0 | not}}
-                ({{.appwebsite}}) 
-              {{end}}
-              would like to perform actions on your behalf, with scope <em>{{.scope}}</em>.
-            </p>
-            <p>The application will redirect to {{.redirect}} to continue.</p>
-            <p>
-                <button
-                    type="submit"
-                    style="width:200px;"
-                >
-                    Allow
-                </button>
-            </p>
-        </form>
-    </main>
-{{ template "footer.tmpl" .}}
+{{- template "header.tmpl" . }}
+<main>
+    <form action="/oauth/authorize" method="POST">
+        <h1>Hi {{ .user -}}!</h1>
+        <p>
+            Application
+            {{- if .appwebsite }}
+            <a href="{{- .appwebsite -}}" rel="nofollow noreferrer noopener" target="_blank">{{- .appname -}}</a>
+            {{- else }}
+            <b>{{- .appname -}}</b>
+            {{- end }}
+            would like to perform actions on your behalf, with scope
+            <em>{{- .scope -}}</em>.
+        </p>
+        <p>
+            To continue, the application will redirect to: <code>{{- .redirect -}}</code>
+        </p>
+        <button type="submit" style="width:200px;">Allow</button>
+    </form>
+</main>
+{{- template "footer.tmpl" . }}
