+++ web/template/confirmed.tmpl
@@ -19,10 +19,10 @@
 
 {{ template "header.tmpl" .}}
 <main>
-	<section>
-		<h1>Email Address Confirmed</h1>
-		<p>Thanks {{.username}}! Your email address <b>{{.email}}</b> has been confirmed.<p>
-	</section>
+    <section>
+        <h1>Email Address Confirmed</h1>
+        <p>Thanks {{.username}}! Your email address <b>{{.email}}</b> has been confirmed.<p>
+    </section>
 </main>
 
 {{ template "footer.tmpl" .}}
