+++ web/template/domain-blocklist.tmpl
@@ -19,34 +19,34 @@
 
 {{ template "header.tmpl" .}}
 <main>
-	<section>
-		<h1>Suspended Instances</h1>
-		<p>
-			The following list of domains have been suspended by the administrator(s) of this server.
-		</p>
-		<p>
-			All current and future accounts on these instances are blocked, and no more data is federated to the remote
-			servers.
-			This extends to subdomains, so an entry for 'example.com' includes 'social.example.com' as well.
-		</p>
-		<div class="list domain-blocklist">
-			<div class="header entry">
-				<div class="domain">Domain</div>
-				<div class="public_comment">Public comment</div>
-			</div>
-			{{range .blocklist}}
-			<div class="entry" id="{{.Domain}}">
-				<div class="domain">
-					<a class="text-cutoff" href="#{{.Domain}}" title="{{.Domain}}">{{.Domain}}</a>
-				</div>
-				<div class="public_comment">
-					<p>
-						{{.PublicComment}}
-					</p>
-				</div>
-			</div>
-			{{end}}
-		</div>
-	</section>
+    <section>
+        <h1>Suspended Instances</h1>
+        <p>
+            The following list of domains have been suspended by the administrator(s) of this server.
+        </p>
+        <p>
+            All current and future accounts on these instances are blocked, and no more data is federated to the remote
+            servers.
+            This extends to subdomains, so an entry for 'example.com' includes 'social.example.com' as well.
+        </p>
+        <div class="list domain-blocklist">
+            <div class="header entry">
+                <div class="domain">Domain</div>
+                <div class="public_comment">Public comment</div>
+            </div>
+            {{range .blocklist}}
+            <div class="entry" id="{{.Domain}}">
+                <div class="domain">
+                    <a class="text-cutoff" href="#{{.Domain}}" title="{{.Domain}}">{{.Domain}}</a>
+                </div>
+                <div class="public_comment">
+                    <p>
+                        {{.PublicComment}}
+                    </p>
+                </div>
+            </div>
+            {{end}}
+        </div>
+    </section>
 </main>
 {{ template "footer.tmpl" .}}
