+++ web/template/frontend.tmpl
@@ -19,7 +19,7 @@
 
 {{ template "header.tmpl" .}}
 <main class="lightgray">
-	<div id="root">
-	</div>
+    <div id="root">
+    </div>
 </main>
 {{ template "footer.tmpl" .}}
