+++ web/template/index.tmpl
@@ -19,59 +19,59 @@
 
 {{ template "header.tmpl" .}}
 <section class="excerpt-top">
-	home to <span class="count">{{.instance.Stats.user_count}}</span> users
-		who posted <span class="count">{{.instance.Stats.status_count}}</span> statuses,
-		federating with  <span class="count">{{.instance.Stats.domain_count}}</span> other instances.
+    home to <span class="count">{{.instance.Stats.user_count}}</span> users
+        who posted <span class="count">{{.instance.Stats.status_count}}</span> statuses,
+        federating with  <span class="count">{{.instance.Stats.domain_count}}</span> other instances.
 </section>
 <main class="lightgray">
-	<section>
-		<div className="short-description">
-			{{.instance.ShortDescription |noescape}}
-		</div>
-	</section>
-	<section class="apps">
-		<p>
-			GoToSocial does not provide its own webclient, but implements the Mastodon client API.
-			You can use this server through a variety of other clients:
-		</p>
-		<div class="applist">
-			<div class="entry">
-				<svg role="img" aria-labelledby="semaphoreTitle semaphoreDesc" class="logo redraw" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 146 120">
-					<title id="semaphoreTitle">The Semaphore logo</title>
-					<desc id="semaphoreDesc">A waving flag</desc>
-					<path d="M68.13 0C53.94 0 42.81 20 13.9 27.1l-2.23-5.29a6.5 6.5 0 0 0-5.17-10.4 6.5 6.5 0 0 0-.81 12.95L46.2 120l5.99-2.5-14.42-33.33c22.8-6.86 32.51-22.16 49.83-20.58 9.9.9 4.87 19.56 8.11 17.93 16.22-8.15 32.44-11.41 50.29-11.41-7.96-9.78-17.38-20.55-22.71-31.74L120.8 32c-2.32-7.33-2.56-14.75.87-22.22-9.74-3.26-21.1 0-32.45 4.9C82.2 9.77 79.5 0 68.13 0zM15.26 30.42c8.95 6.63 13.63 13.86 16.07 20.94l1.62 6.32c1.24 6.58 1.07 12.8 1.27 18.03z"></path>
-				</svg>
-				<div>
-					<h2>Semaphore</h2>
-					<p>Semaphore is a web client designed for speed and simplicity.</p>
-					<a href="https://semaphore.social/" target="_blank" rel="noopener">Use Semaphore</a>
-				</div>
-			</div>
-			<div class="entry">
-				<img class="logo" src="/assets/tusky.svg" alt="The Tusky mascot, a cartoon elephant tooting happily"/>
-				<div>
-					<h2>Tusky</h2>
-					<p>Tusky is a lightweight mobile client for Android.</p>
-					<a href="https://tusky.app" target="_blank" rel="noopener">Get Tusky</a>
-				</div>
-			</div>
-			<div class="entry">
-				<img class="logo" src="/assets/feditext.svg" alt="The Feditext logo, the characters ft at a slight angle">
-				<div>
-					<h2>Feditext</h2>
-					<p>Feditext (beta) is a beautiful client for iOS, iPadOS and macOS.</p>
-					<a href="https://fedi.software/@Feditext" target="_blank" rel="noopener">Get Feditext</a>
-				</div>
-			</div>
-			<div class="entry">
-				<img class="logo" src="/assets/mastodon.svg" alt="The Mastodon logo, the character M in a speech bubble">
-				<div>
-					<h2>More clients</h2>
-					<p>Or try one of the clients listed on the official Mastodon page.</p>
-					<a href="https://joinmastodon.org/apps" target="_blank" rel="noopener">Get Mastodon apps</a>
-				</div>
-			</div>
-		</div>
-	</section>
+    <section>
+        <div className="short-description">
+            {{.instance.ShortDescription |noescape}}
+        </div>
+    </section>
+    <section class="apps">
+        <p>
+            GoToSocial does not provide its own webclient, but implements the Mastodon client API.
+            You can use this server through a variety of other clients:
+        </p>
+        <div class="applist">
+            <div class="entry">
+                <svg role="img" aria-labelledby="semaphoreTitle semaphoreDesc" class="logo redraw" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 146 120">
+                    <title id="semaphoreTitle">The Semaphore logo</title>
+                    <desc id="semaphoreDesc">A waving flag</desc>
+                    <path d="M68.13 0C53.94 0 42.81 20 13.9 27.1l-2.23-5.29a6.5 6.5 0 0 0-5.17-10.4 6.5 6.5 0 0 0-.81 12.95L46.2 120l5.99-2.5-14.42-33.33c22.8-6.86 32.51-22.16 49.83-20.58 9.9.9 4.87 19.56 8.11 17.93 16.22-8.15 32.44-11.41 50.29-11.41-7.96-9.78-17.38-20.55-22.71-31.74L120.8 32c-2.32-7.33-2.56-14.75.87-22.22-9.74-3.26-21.1 0-32.45 4.9C82.2 9.77 79.5 0 68.13 0zM15.26 30.42c8.95 6.63 13.63 13.86 16.07 20.94l1.62 6.32c1.24 6.58 1.07 12.8 1.27 18.03z"></path>
+                </svg>
+                <div>
+                    <h2>Semaphore</h2>
+                    <p>Semaphore is a web client designed for speed and simplicity.</p>
+                    <a href="https://semaphore.social/" target="_blank" rel="noopener">Use Semaphore</a>
+                </div>
+            </div>
+            <div class="entry">
+                <img class="logo" src="/assets/tusky.svg" alt="The Tusky mascot, a cartoon elephant tooting happily"/>
+                <div>
+                    <h2>Tusky</h2>
+                    <p>Tusky is a lightweight mobile client for Android.</p>
+                    <a href="https://tusky.app" target="_blank" rel="noopener">Get Tusky</a>
+                </div>
+            </div>
+            <div class="entry">
+                <img class="logo" src="/assets/feditext.svg" alt="The Feditext logo, the characters ft at a slight angle">
+                <div>
+                    <h2>Feditext</h2>
+                    <p>Feditext (beta) is a beautiful client for iOS, iPadOS and macOS.</p>
+                    <a href="https://fedi.software/@Feditext" target="_blank" rel="noopener">Get Feditext</a>
+                </div>
+            </div>
+            <div class="entry">
+                <img class="logo" src="/assets/mastodon.svg" alt="The Mastodon logo, the character M in a speech bubble">
+                <div>
+                    <h2>More clients</h2>
+                    <p>Or try one of the clients listed on the official Mastodon page.</p>
+                    <a href="https://joinmastodon.org/apps" target="_blank" rel="noopener">Get Mastodon apps</a>
+                </div>
+            </div>
+        </div>
+    </section>
 </main>
 {{ template "footer.tmpl" .}}
