+++ web/template/oob.tmpl
@@ -19,10 +19,10 @@
 
 {{ template "header.tmpl" .}}
 <main>
-	<section class="oob-token">
-		<h1>Hi {{ .user }}!</h1>
-		<p>Here's your out-of-band token with scope "<em>{{.scope}}</em>", use it wisely:</p>
-		<code>{{ .oobToken }}</code>
-	</section>
+    <section class="oob-token">
+        <h1>Hi {{ .user }}!</h1>
+        <p>Here's your out-of-band token with scope "<em>{{.scope}}</em>", use it wisely:</p>
+        <code>{{ .oobToken }}</code>
+    </section>
 </main>
 {{ template "footer.tmpl" .}}
