+++ web/template/status.tmpl
@@ -19,78 +19,78 @@
 
 <a data-nosnippet href="{{- .URL -}}" class="toot-link">Open thread</a>
 <section class="author">
-	<a href="{{- .Account.URL -}}">
-		<img class="avatar" src="{{- .Account.Avatar -}}" alt="">
-		<span aria-hidden="true" class="displayname">
-			{{- if .Account.DisplayName -}}
-			{{- emojify .Account.Emojis (escape .Account.DisplayName) -}}
-			{{- else -}}
-			{{- .Account.Username -}}
-			{{- end -}}
-		</span>
-		<span aria-hidden="true" class="username">@{{- .Account.Username -}}</span>
-		<span class="sr-only">
-			{{- if .Account.DisplayName -}}
-			{{- emojify .Account.Emojis (escape .Account.DisplayName) -}}. Username: @{{ .Account.Acct -}}.
-			{{- else -}}
-			@{{- .Account.Acct -}}.
-			{{- end -}}
-		</span>
-	</a>
+    <a href="{{- .Account.URL -}}">
+        <img class="avatar" src="{{- .Account.Avatar -}}" alt="">
+        <span aria-hidden="true" class="displayname">
+            {{- if .Account.DisplayName -}}
+            {{- emojify .Account.Emojis (escape .Account.DisplayName) -}}
+            {{- else -}}
+            {{- .Account.Username -}}
+            {{- end -}}
+        </span>
+        <span aria-hidden="true" class="username">@{{- .Account.Username -}}</span>
+        <span class="sr-only">
+            {{- if .Account.DisplayName -}}
+            {{- emojify .Account.Emojis (escape .Account.DisplayName) -}}. Username: @{{ .Account.Acct -}}.
+            {{- else -}}
+            @{{- .Account.Acct -}}.
+            {{- end -}}
+        </span>
+    </a>
 </section>
 <section class="body">
-	{{- if .SpoilerText }}
-	<details class="text-spoiler">
-		<summary>
-			<span class="spoiler-text" lang="{{- .LanguageTag.TagStr -}}">{{- emojify .Emojis (escape .SpoilerText) -}}</span>
-			<span class="button" role="button" tabindex="0">Toggle visibility</span>
-		</summary>
-		<div class="text">
-			<div class="content" lang="{{- .LanguageTag.TagStr -}}">
-				{{ emojify .Emojis (noescape .Content) }}
-			</div>
-			{{- if .Poll }}
-			{{ template "status_poll.tmpl" . }}
-			{{- end }}
-		</div>
-	</details>
-	{{- else }}
-	<div class="text">
-		<div class="content" lang="{{- .LanguageTag.TagStr -}}">
-			{{ emojify .Emojis (noescape .Content) }}
-		</div>
-		{{- if .Poll }}
-		{{ template "status_poll.tmpl" . }}
-		{{- end }}
-	</div>
-	{{- end }}
-	{{- if .MediaAttachments }}
-	{{ template "status_attachments.tmpl" . }}
-	{{- end }}
+    {{- if .SpoilerText }}
+    <details class="text-spoiler">
+        <summary>
+            <span class="spoiler-text" lang="{{- .LanguageTag.TagStr -}}">{{- emojify .Emojis (escape .SpoilerText) -}}</span>
+            <span class="button" role="button" tabindex="0">Toggle visibility</span>
+        </summary>
+        <div class="text">
+            <div class="content" lang="{{- .LanguageTag.TagStr -}}">
+                {{ emojify .Emojis (noescape .Content) }}
+            </div>
+            {{- if .Poll }}
+            {{ template "status_poll.tmpl" . }}
+            {{- end }}
+        </div>
+    </details>
+    {{- else }}
+    <div class="text">
+        <div class="content" lang="{{- .LanguageTag.TagStr -}}">
+            {{ emojify .Emojis (noescape .Content) }}
+        </div>
+        {{- if .Poll }}
+        {{ template "status_poll.tmpl" . }}
+        {{- end }}
+    </div>
+    {{- end }}
+    {{- if .MediaAttachments }}
+    {{ template "status_attachments.tmpl" . }}
+    {{- end }}
 </section>
 <aside class="info">
-	<time datetime="{{- .CreatedAt -}}">{{- .CreatedAt | timestampPrecise -}}</time>
-	<div class="stats" role="group">
-		<div class="stats-item">
-			<span aria-hidden="true"><i class="fa fa-reply-all"></i> {{ .RepliesCount -}}</span>
-			<span class="sr-only">{{- .RepliesCount }} {{ if .RepliesCount | eq 1 }}reply{{ else }}replies{{ end -}}</span>
-		</div>
-		<div class="stats-item">
-			<span aria-hidden="true"><i class="fa fa-star"></i> {{ .FavouritesCount -}}</span>
-			<span class="sr-only">{{- .FavouritesCount }} {{ if .FavouritesCount | eq 1 }}favourite{{ else }}favourites{{ end -}}</span>
-		</div>
-		<div class="stats-item">
-			<span aria-hidden="true"><i class="fa fa-retweet"></i> {{ .ReblogsCount -}}</span>
-			<span class="sr-only">{{- .ReblogsCount }} {{ if .ReblogsCount | eq 1 }}boost{{ else }}boosts{{ end -}}</span>
-		</div>
-		{{- if .Pinned }}
-		<div class="stats-item">
-			<i class="fa fa-thumb-tack" aria-hidden="true"></i>
-			<span class="sr-only">pinned</span>
-		</div>
-		{{- end }}
-		{{- if .LanguageTag.DisplayStr }}
-		<div class="stats-item language" title="Language: {{ .LanguageTag.DisplayStr }}">{{ .LanguageTag.TagStr }}</div>
-		{{- end }}
-	</div>
+    <time datetime="{{- .CreatedAt -}}">{{- .CreatedAt | timestampPrecise -}}</time>
+    <div class="stats" role="group">
+        <div class="stats-item">
+            <span aria-hidden="true"><i class="fa fa-reply-all"></i> {{ .RepliesCount -}}</span>
+            <span class="sr-only">{{- .RepliesCount }} {{ if .RepliesCount | eq 1 }}reply{{ else }}replies{{ end -}}</span>
+        </div>
+        <div class="stats-item">
+            <span aria-hidden="true"><i class="fa fa-star"></i> {{ .FavouritesCount -}}</span>
+            <span class="sr-only">{{- .FavouritesCount }} {{ if .FavouritesCount | eq 1 }}favourite{{ else }}favourites{{ end -}}</span>
+        </div>
+        <div class="stats-item">
+            <span aria-hidden="true"><i class="fa fa-retweet"></i> {{ .ReblogsCount -}}</span>
+            <span class="sr-only">{{- .ReblogsCount }} {{ if .ReblogsCount | eq 1 }}boost{{ else }}boosts{{ end -}}</span>
+        </div>
+        {{- if .Pinned }}
+        <div class="stats-item">
+            <i class="fa fa-thumb-tack" aria-hidden="true"></i>
+            <span class="sr-only">pinned</span>
+        </div>
+        {{- end }}
+        {{- if .LanguageTag.DisplayStr }}
+        <div class="stats-item language" title="Language: {{ .LanguageTag.DisplayStr }}">{{ .LanguageTag.TagStr }}</div>
+        {{- end }}
+    </div>
 </aside>
