+++ web/template/status_poll.tmpl
@@ -18,45 +18,45 @@
 */ -}}
 
 {{- /*
-		Template for rendering a web view of a poll.
-		To use this template, pass a web view status into it.
+        Template for rendering a web view of a poll.
+        To use this template, pass a web view status into it.
 */ -}}
 
-	<figure class="poll">
-		<figcaption class="poll-info">
-			<span class="poll-expiry">
-				{{- if .Poll.Multiple -}}
-				Multiple-choice poll&nbsp;
-				{{- else -}}
-				Poll&nbsp;
-				{{- end -}}
-				{{- if .Poll.Expired -}}
-				closed {{ .Poll.ExpiresAt | timestampPrecise -}}
-				{{- else if .Poll.ExpiresAt -}}
-				open until {{ .Poll.ExpiresAt | timestampPrecise -}}
-				{{- else -}}
-				open forever
-				{{- end -}}
-			</span>
-			<span class="total-votes">Total votes: {{ .Poll.VotesCount }}</span>
-		</figcaption>
-		<ul class="poll-options">
-		{{- range $index, $pollOption := .WebPollOptions }}
-			<li class="poll-option">
-				<label aria-hidden="true" for="poll-{{- $pollOption.PollID -}}-option-{{- increment $index -}}" lang="{{- .LanguageTag.TagStr -}}">{{- emojify .Emojis (noescape $pollOption.Title) -}}</label>
-				<meter aria-hidden="true" id="poll-{{- $pollOption.PollID -}}-option-{{- increment $index -}}" min="0" max="100" value="{{- $pollOption.VoteShare -}}">{{- $pollOption.VoteShare -}}&#37;</meter>
-				<div class="sr-only">Option {{ increment $index }}:&nbsp;<span lang="{{ .LanguageTag.TagStr }}">{{ emojify .Emojis (noescape $pollOption.Title) -}}</span></div>
-				<div class="poll-vote-summary">
-					<span class="poll-vote-share">{{- $pollOption.VoteShareStr -}}&#37;</span>
-					<span class="poll-vote-count">
-						{{- if eq $pollOption.VotesCount 1 -}}
-							{{- $pollOption.VotesCount }} vote
-						{{- else -}}
-							{{- $pollOption.VotesCount }} votes
-						{{- end -}}
-					</span>
-				</div>
-			</li>
-		{{- end }}
-		</ul>
-	</figure>
+    <figure class="poll">
+        <figcaption class="poll-info">
+            <span class="poll-expiry">
+                {{- if .Poll.Multiple -}}
+                Multiple-choice poll&nbsp;
+                {{- else -}}
+                Poll&nbsp;
+                {{- end -}}
+                {{- if .Poll.Expired -}}
+                closed {{ .Poll.ExpiresAt | timestampPrecise -}}
+                {{- else if .Poll.ExpiresAt -}}
+                open until {{ .Poll.ExpiresAt | timestampPrecise -}}
+                {{- else -}}
+                open forever
+                {{- end -}}
+            </span>
+            <span class="total-votes">Total votes: {{ .Poll.VotesCount }}</span>
+        </figcaption>
+        <ul class="poll-options">
+        {{- range $index, $pollOption := .WebPollOptions }}
+            <li class="poll-option">
+                <label aria-hidden="true" for="poll-{{- $pollOption.PollID -}}-option-{{- increment $index -}}" lang="{{- .LanguageTag.TagStr -}}">{{- emojify .Emojis (noescape $pollOption.Title) -}}</label>
+                <meter aria-hidden="true" id="poll-{{- $pollOption.PollID -}}-option-{{- increment $index -}}" min="0" max="100" value="{{- $pollOption.VoteShare -}}">{{- $pollOption.VoteShare -}}&#37;</meter>
+                <div class="sr-only">Option {{ increment $index }}:&nbsp;<span lang="{{ .LanguageTag.TagStr }}">{{ emojify .Emojis (noescape $pollOption.Title) -}}</span></div>
+                <div class="poll-vote-summary">
+                    <span class="poll-vote-share">{{- $pollOption.VoteShareStr -}}&#37;</span>
+                    <span class="poll-vote-count">
+                        {{- if eq $pollOption.VotesCount 1 -}}
+                            {{- $pollOption.VotesCount }} vote
+                        {{- else -}}
+                            {{- $pollOption.VotesCount }} votes
+                        {{- end -}}
+                    </span>
+                </div>
+            </li>
+        {{- end }}
+        </ul>
+    </figure>
