+++ web/template/tag.tmpl
@@ -20,8 +20,8 @@
 {{ template "header.tmpl" .}}
 
 <main class="thread">
-	<h2 id="tag-name" tabindex="-1">#{{.tagName}}</h2>
-	<p>There's nothing here yet!</p>
+    <h2 id="tag-name" tabindex="-1">#{{.tagName}}</h2>
+    <p>There's nothing here yet!</p>
 </main>
 
 {{ template "footer.tmpl" .}}
