+++ web/template/thread.tmpl
@@ -19,20 +19,20 @@
 
 {{ template "header.tmpl" .}}
 <main>
-	<section data-nosnippet class="thread">
-		{{range .context.Ancestors}}
-		<article class="toot" id="{{.ID}}">
-			{{ template "status.tmpl" .}}
-		</article>
-		{{end}}
-		<article class="toot expanded" id="{{.status.ID}}">
-			{{ template "status.tmpl" .status}}
-		</article>
-		{{range .context.Descendants}}
-		<article class="toot" id="{{.ID}}">
-			{{ template "status.tmpl" .}}
-		</article>
-		{{end}}
-	</section>
+    <section data-nosnippet class="thread">
+        {{range .context.Ancestors}}
+        <article class="toot" id="{{.ID}}">
+            {{ template "status.tmpl" .}}
+        </article>
+        {{end}}
+        <article class="toot expanded" id="{{.status.ID}}">
+            {{ template "status.tmpl" .status}}
+        </article>
+        {{range .context.Descendants}}
+        <article class="toot" id="{{.ID}}">
+            {{ template "status.tmpl" .}}
+        </article>
+        {{end}}
+    </section>
 </main>
 {{ template "footer.tmpl" .}}
